+++ docs/CHANGELOG.md
@@ -5,6 +5,13 @@
 - Package `ml/activations`
   - Added `HardSigmoid` activation.
   - Modified the parametrized activations to be suffixed with `With`: `LeakyReluWith`, `HardSigmoidWith`, `HardSwishWith`.
+- Package `attention`:
+  - API clean up.
+  - Dropout now takes a `*Node` instead of a float64, allowing dynamic dropout control.
+- Package `attention/pos`:
+  - Split `Encoder` API into `QKEncoder` and `PreEncoder`, to support different types of positional encoders.
+- Pacakge `ml/models/transfromers`:
+  - Fixes and clean ups of the API; Updated positional-encoder support.
 - Package `tensors`
   - Improving support for sub-byte data types (`Int4`, `Int2`, `Uint4`, `Uint2`)